Create CoworkersUpdate.vue.

+++ src/pages/Coworkers/CoworkersUpdate.vue
@@ -1,54 +1,56 @@
 <template>
   <div class="coworkers-update">
-    <div class="coworkers-update__header">
-      <h1 class="coworkers-update__heading">{{coworker.name}}</h1>
-      <div class="coworkers-update__button">
-        <button-unit
-            class="coworkers-update__button-delete"
-            color="red"
-            @click="handleDelete"
-        >
-          Delete
-        </button-unit>
-        <button-unit
-            class="coworkers-update__button-edit"
-            color="purple"
-            @click="handleEdit"
-        >
-          Edit
-        </button-unit>
+    <Form
+        class="coworkers-update__form"
+        @submit="onSubmit"
+        :validation-schema="schema"
+    >
+      <div class="coworkers-update__header">
+        <h1 class="coworkers-update__heading">{{coworker.name}}</h1>
+        <div class="coworkers-update__button">
+          <button-unit
+              class="coworkers-update__button-delete"
+              color="red"
+              @click="handleDelete"
+          >
+            Delete
+          </button-unit>
+          <button-unit
+              class="coworkers-update__button-edit"
+              color="purple"
+              type="submit"
+          >
+            Edit
+          </button-unit>
+        </div>
       </div>
-    </div>
-    <form class="coworkers-update__form">
       <label
           class="coworkers-update__label"
           for="name"
       >
         Name
       </label>
-      <input
+      <Field
           class="coworkers-update__input"
           type="text"
           id="name"
           name="name"
-          placeholder="name"
-          v-model="coworker.name"
       />
+      <ErrorMessage name="name" />
       <label
           class="coworkers-update__label"
           for="password"
       >
         Password
       </label>
-      <input
+      <Field
           class="coworkers-update__input"
           type="password"
           id="password"
           name="password"
-          placeholder="password"
-          v-model="coworker.password"
       />
-    </form>
+      <ErrorMessage name="password" />
+    </Form>
   </div>
 </template>
 
@@ -57,10 +59,12 @@ import {defineComponent, ref} from "vue";
 import {useRouter} from "vue-router";
 import axios from "axios";
 import ButtonUnit from "@/components/atoms/ButtonUnit/ButtonUnit.vue";
+import { Field, Form, ErrorMessage } from 'vee-validate';
+import * as yup from 'yup';
 
 export default defineComponent({
   name: 'CoworkersUpdate',
-  components: {ButtonUnit},
+  components: {ButtonUnit, Field, Form, ErrorMessage},
   setup() {
     const router = useRouter();
     const id: number = Number(router.currentRoute.value.params.id);
@@ -75,16 +79,27 @@ export default defineComponent({
       coworker
     }
   },
+  data() {
+    const schema = yup.object({
+      name: yup.string().required().label('Name'),
+      password: yup.string().required().min(8).label('Password'),
+    });
+    return {
+      schema,
+    };
+  },
   methods: {
-    handleEdit(): void {
-      axios.post(`${process.env.VUE_APP_BACKEND_URL}/coworkers/${this.id}`, this.coworker);
+    onSubmit(values: any) {
+      axios.post(`${process.env.VUE_APP_BACKEND_URL}/coworkers/${this.id}`, values);
       this.$router.push({ name: 'CoworkersRetrieveAll' });
     },
     handleDelete(): void {
-      axios.delete(`${process.env.VUE_APP_BACKEND_URL}/coworkers/${this.id}`);
-      this.$router.push({ name: 'CoworkersRetrieveAll' });
+      if (confirm("Do you want to delete?")) {
+        axios.delete(`${process.env.VUE_APP_BACKEND_URL}/coworkers/${this.id}`);
+        this.$router.push({ name: 'CoworkersRetrieveAll' });
+      }
     },
-  }
+  },
 })
 </script>
